% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDT.R
\docType{data}
\name{DDT}
\alias{DDT}
\title{DDT Dataset}
\format{
A data frame with 144 rows and 6 variables:
\describe{
  \item{RIVER}{River mouths}
  \item{MILE}{Miles along the rivers}
  \item{SPECIES}{Species which was tested}
  \item{LENGTH}{the fishes length}
  \item{WEIGHT}{the fished weight}
  \item{DDT}{Amount of DDT in the fish}
}
}
\source{
Original CSV file stored in \code{inst/extdata/DDT.csv}.
}
\usage{
data(DDT)
}
\description{
A dataset containing 144 fish with there Species,Length,Weight,what mile they were located, the river they were located, and their DDT level
}
\examples{
data(DDT)
head(DDT)
}
\keyword{datasets}
