% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Data}
\format{
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{country}{Country name}
  \item{iso2}{2-letter ISO country code}
  \item{iso3}{3-letter ISO country code}
  \item{year}{Year of observation}
  \item{...}{Other variables from the WHO dataset}
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
fire
}
\description{
A subset of data from the World Health Organization Global Tuberculosis Report.
}
\keyword{datasets}
