% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addingfunc.R
\name{myread}
\alias{myread}
\title{my read function}
\usage{
myread(csv)
}
\arguments{
\item{csv}{the data file}
}
\value{
a table from the csv file
}
\description{
my read function
}
\examples{
myread(DDT.csv)
}
