% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.r
\name{pareto}
\alias{pareto}
\title{Pareto Barplot}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{vector data set}

\item{mn}{title}

\item{...}{}
}
\value{
a pareto barplot of a given data set
}
\description{
Pareto Barplot
}
\examples{
freq <- c(15,8,63,20);RL <- c("None","Both","LegsO","WheelsO");l <- rep(RL, freq);pareto(l)
}
