% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypergeometric.R
\name{myhyper}
\alias{myhyper}
\title{Hypergeometric function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{iterations through parameter values}

\item{N}{population size}

\item{r}{Successes in population}

\item{n}{sample size}
}
\value{
a hypergeometric simulation and plot
}
\description{
Hypergeometric function
}
\examples{
myhyper(iter= 100,N=20,r=12,n=5)
}
